Osmi commit - Prvi domaci

+++ app/Http/Controllers/ContactController.php
@@ -24,15 +24,13 @@ class ContactController extends Controller
         $request->validate([
             'email' => 'required|string',
             'subject' => 'required|string',
-            'description' => 'required|string|min:5' // description mora biti sa minimum 5 slova
+            'message' => 'required|string|min:5', // message mora biti sa minimum 5 karaktera
         ]);
-
-        echo "Email: ".$request->get('email'). " Naslov: ".$request->get('subject'). " Poruka: ".$request->get('description');
 
         Contact::create([
             'email' => $request->get('email'),
             'subject' => $request->get('subject'),
-            'message' => $request->get('description')
+            'message' => $request->get('message'),
         ]);
 
         return redirect('/shop');
@@ -41,8 +39,35 @@ class ContactController extends Controller
     public function delete($contact)
     {
         $singleContact = Contact::where(['id' => $contact])->first();
+
+        if ($singleContact === null){
+            return redirect()->back()->with('message', 'Kontakt ne postoji!');
+        }
         $singleContact->delete();
 
-        return redirect()->back();
+        return redirect()->route('adminAllContacts');
+    }
+
+    public function updateContactForm($contact)
+    {
+        $singleContact = Contact::findOrFail($contact);
+
+        return view('updateContactForm', compact('singleContact'));
+    }
+
+    public function update(Request $request, Contact $contact)
+    {
+        $request->validate([
+            'email' => 'required|string',
+            'subject' => 'required|string',
+            'message' => 'required|string|min:5',
+        ]);
+        $contact->update([
+            'email' => $request->get('email'),
+            'subject' => $request->get('subject'),
+            'message' => $request->get('message'),
+        ]);
+
+        return redirect()->route('adminAllContacts');
     }
 }
